refactor(devnet): move apis to its own modules
typo

fmt

fix unsed
